requirements.txt
# General RAG & Bot deps
llama-index>=0.10.0
llama-index-llms-google-genai>=0.1.0
llama-index-embeddings-huggingface>=0.2.0
sentence-transformers>=2.2.0
google-genai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0

# Frontend API & VectorDB deps
psycopg2-binary>=2.9.0
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# GraphRAG deps
# Using built-in SimplePropertyGraphStore for local JSON persistence
